Refactor e2e tests to separate common helper functions
